+++ tests/integration/test_assets_service.py
